+++ src/pages/tasks/newTaskForm.tsx
@@ -1,4 +1,4 @@
-// import { Link } from "react-router-dom";
+import { Link } from "react-router-dom";
 import Button from "@mui/material/Button";
 import Navbar from "../../components/navbar";
 import Sidebar from "../../components/sidebar";
@@ -8,10 +8,10 @@ import { useNavigate } from "react-router-dom";
 import Box from "@mui/material/Box";
 import TextField from "@mui/material/TextField";
 import MenuItem from "@mui/material/MenuItem";
-// import Typography from "@mui/material/Typography";
+import Typography from "@mui/material/Typography";
 import { useState } from "react";
 import { Project } from "../../types/project";
-// import DatesButton from "../tasks/datesButtons";
+// import DatesButton from "../timeTracking/datesButtons";
 
 import axios from "axios";
 axios.defaults.withCredentials = true;
@@ -115,11 +115,7 @@ const NewTaskFrom = () => {
     <>
       <Navbar />
       <Sidebar />
-      {/* <DatesButton
-        dates={dates}
-        selectedDate={selectedDate}
-        setSelectedDate={setSelectedDate}
-      /> */}
+
       <Box style={{ width: "80%", marginLeft: "32%", marginTop: "90px" }}>
         <h3>New Task</h3>
         <form onSubmit={(e: React.SyntheticEvent) => handleGetProject(e)}>
