+++ src/pages/tasks/task.tsx
@@ -7,7 +7,7 @@ import { DateTime } from "luxon";
 import { Spinner } from "../../components/spinner/spinner";
 import Typography from "@mui/material/Typography";
 import Box from "@mui/material/Box";
-import DatesButton from "../tasks/datesButtons";
+import DatesButton from "../timeTracking/datesButtons";
 import { Link } from "react-router-dom";
 import Button from "@mui/material/Button";
 import { useNavigate } from "react-router-dom";
